requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
anytree==2.12.1
appdirs==1.4.4
ascii_colors==0.5.2
async-timeout==4.0.3
attrs==24.2.0
autograd==1.7.0
babel==2.17.0
backrefs==5.9
beartype==0.18.5
blis==1.2.0
bottle==0.13.2
catalogue==2.0.10
cefpython3==66.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.20.0
clr-loader==0.2.6
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.12
confection==0.1.5
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
distro==1.9.0
elevenlabs==1.10.0
elevenlabslib==0.31.3
etelemetry==0.3.1
exceptiongroup==1.2.2
faiss-cpu==1.10.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.56.0
fr_core_news_md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.8.0/fr_core_news_md-3.8.0-py3-none-any.whl#sha256=8a70d090a54ef77525c3ffa6a6195b9d365f2cf369ae1cd84ede93f3d709079e
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
fuzzywuzzy==0.18.0
gensim==4.3.3
geojson==2.5.0
ghp-import==2.1.0
graspologic==3.4.1
graspologic-native==1.2.3
greenlet==3.1.1
groq==0.31.0
h11==0.14.0
hid==1.0.7
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.1
huggingface-hub==0.26.1
humanfriendly==10.0
hyppo==0.4.0
idna==3.10
importlib_resources==6.4.5
isodate==0.6.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-groq==0.3.6
langchain-openai==0.3.6
langchain-text-splitters==0.3.9
langcodes==3.5.0
langsmith==0.4.8
language_data==1.3.0
Levenshtein==0.26.0
llvmlite==0.43.0
looseversion==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.10.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.20
mkdocs-material-extensions==1.3.1
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
nano-vectordb==0.0.4.3
networkx==3.4.2
nibabel==5.3.2
nipype==1.8.6
numba==0.60.0
numpy==1.26.4
onnxruntime==1.19.2
openai==1.63.2
orjson==3.10.10
packaging==24.1
paginate==0.5.7
pandas==2.2.3
pathspec==0.12.1
patsy==1.0.1
pefile==2023.2.7
pillow==11.0.0
pipdeptree==2.24.0
pipmaster==0.4.0
platformdirs==4.4.0
pluggy==1.5.0
POT==0.9.5
preshed==3.0.9
propcache==0.2.0
protobuf==5.28.2
prov==2.0.1
proxy-tools==0.1.0
psutil==6.1.0
PyAudio==0.2.14
pybind11==2.13.6
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydot==3.0.2
pydub==0.25.1
Pygments==2.19.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pymdown-extensions==10.16.1
PyMuPDF==1.24.12
pymupdf4llm==0.0.17
pynndescent==0.5.13
pyowm==3.3.0
pyparsing==3.2.0
pypdf==5.1.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.26.0
pythonnet==3.0.4
pytz==2024.2
pywebview==5.3
pywin32==308
pywin32-ctypes==0.2.3
pywinauto==0.6.9
pywinusb==0.4.2
pywttr-models==2.0.1
pyxnat==1.6.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
RapidFuzz==3.10.0
rdflib==6.3.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
resampy==0.4.3
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==3.2.1
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechify-api==1.2.0
SQLAlchemy==2.0.36
srsly==2.5.1
srt==3.5.3
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
traits==6.3.2
transformers==4.46.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.2.3
vosk==0.3.45
vue==0.0.1
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
webrtcvad-wheels==2.0.14
websockets==13.1
winrt-runtime==3.2.1
wrapt==1.17.2
xmltodict==0.14.2
xx_ent_wiki_sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.8.0/xx_ent_wiki_sm-3.8.0-py3-none-any.whl#sha256=6f3c4b853852ea9e9d2dc76cc950dddb10a7e4c42d813308caefe6c5e8be2f0a
yarl==1.16.0
zstandard==0.23.0
